// ui/src/main/angular/ui/modules/lipidcompass-backend-client/model/pagePlottableLipidQuantity.ts
/**
 * LipidCompass Backend API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { PageableObject } from './pageableObject';
import { SortObject } from './sortObject';
import { PlottableLipidQuantity } from './plottableLipidQuantity';


export interface PagePlottableLipidQuantity { 
    totalElements?: number;
    totalPages?: number;
    size?: number;
    content?: Array<PlottableLipidQuantity>;
    number?: number;
    sort?: SortObject;
    pageable?: PageableObject;
    first?: boolean;
    last?: boolean;
    numberOfElements?: number;
    empty?: boolean;
}